+++ CHANGES.md
@@ -3,7 +3,13 @@
 ### ? - ?
 
 ##### Additions :tada:
+
+- Added `CesiumUrlTemplateRasterOverlay`, allowing a raster overlay to be added using tiles requested based on a specified URL template.
 - Added `RequestHeaders` property to `Cesium3DTileset`, allowing per-tileset headers to be specified.
+
+##### Fixes :wrench:
+
+- Fixed another bug in `CesiumSubLevelSwitcherComponent` that could prevent all sub-levels from loading if a single sub-level failed to load.
 
 ### v2.12.0 - 2025-01-02